Moved validity checks into util so that they can be swapped out going forward. Fixed the warning on compile for the util class that i was not used by switching to a while loop. Totally re-ordered attribute to use an uncle bob approach to dependant private methods.
