+++ src/parser/element.rs
@@ -1,14 +1,9 @@
 
 extern crate xml;
 use parser::attribute::*;
-use parser::util::*;
-
-fn valid_react_dom_element (test_element: &str) -> bool {
-    let valid_react_elements = ["circle", "clipPath", "defs", "ellipse", "g", "line", 
-        "linearGradient", "mask", "path", "pattern", "polygon", "polyline", "radialGradient",
-        "rect", "stop", "svg", "text", "tspan"];
-    return valid_react_elements.contains(&test_element)
-}
+use parser::util::tab_in;
+use parser::util::parse_off_extra_w3c_details;
+use parser::util::valid_react_dom_element;
 
 pub fn build_element(name: xml::name::OwnedName,
                  attributes: Vec<xml::attribute::OwnedAttribute>,
